lr4-digital-signature/requirements.txt
asn1==2.7.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
coverage==7.6.7
cycler==0.12.1
dsmltf==0.1.0
enum-compat==0.0.3
execnet==2.1.1
fonttools==4.55.0
idna==3.10
iniconfig==2.0.0
kiwisolver==1.4.7
ksilorama==0.4.10
matplotlib==3.9.2
mpmath==1.3.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
pycryptodome==3.21.0
pyparsing==3.2.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
soupsieve==2.6
sympy==1.13.3
tzdata==2024.2
urllib3==2.2.3
